rebased to latest upstream

+++ rsync_tmbackup.sh
@@ -33,7 +33,10 @@ trap 'fn_terminate_script' SIGINT
 
 fn_parse_date() {
 	# Converts YYYY-MM-DD-HHMMSS to YYYY-MM-DD HH:MM:SS and then to Unix Epoch.
-	date -d "${1:0:10} ${1:11:2}:${1:13:2}:${1:15:2}" +%s
+	case "$OSTYPE" in
+		linux*) date -d "${1:0:10} ${1:11:2}:${1:13:2}:${1:15:2}" +%s ;;
+		darwin*) date -j -f "%Y-%m-%d-%H%M%S" "$1" "+%s" ;;
+	esac
 }
 
 fn_find_backups() {
@@ -41,8 +44,15 @@ fn_find_backups() {
 }
 
 fn_expire_backup() {
-	TMP_EMPTY_DIR=$(mktemp -d /tmp/empty.XXXXXX)
+	# Double-check that we're on a backup destination to be completely
+	# sure we're deleting the right folder
+	if [ -z "$(fn_is_backup_destination "$(dirname -- "$1")")" ]; then
+		fn_log_error "$1 is not on a backup destination - aborting."
+		exit 1
+	fi
+
 	fn_log_info "Expiring $1"
+	TMP_EMPTY_DIR=$(mktemp -d /tmp/empty.XXXXXX)
 	rsync -a --delete $TMP_EMPTY_DIR/ "$1"
   rmdir $TMP_EMPTY_DIR
 }
@@ -56,8 +66,8 @@ DEST_FOLDER=${2%/}
 INCLUSION_FILE=$3
 EXCLUSION_FILE=$4
 
-for arg in "$SRC_FOLDER" "$DEST_FOLDER" "$INCLUSION_FILE" "$EXCLUSION_FILE"; do
-	if [[ "$arg" == *"'"* ]]; then
+for ARG in "$SRC_FOLDER" "$DEST_FOLDER" "$INCLUSION_FILE" "$EXCLUSION_FILE"; do
+if [[ "$ARG" == *"'"* ]]; then
 		fn_log_error 'Arguments may not have any single quote characters.'
 		exit 1
 	fi
@@ -74,15 +84,10 @@ fn_backup_marker_path() {
 }
 
 fn_is_backup_destination() {
-	DEST_MARKER_FILE="$(fn_backup_marker_path $1)"
-	if [ -f "$DEST_MARKER_FILE" ]; then
-		echo "1"
-	else
-		echo "0"
-	fi
+	find "$(fn_backup_marker_path "$1")" 2>/dev/null
 }
 
-if [ "$(fn_is_backup_destination $DEST_FOLDER)" != "1" ]; then
+if [ -z "$(fn_is_backup_destination $DEST_FOLDER)" ]; then
 	fn_log_info "Safety check failed - the destination does not appear to be a backup folder or drive (marker file not found)."
 	fn_log_info "If it is indeed a backup folder, you may add the marker file by running the following command:"
 	fn_log_info ""
@@ -95,15 +100,18 @@ fi
 # Setup additional variables
 # -----------------------------------------------------------------------------
 
-export IFS=$'\n' # Better for handling spaces in filenames.
+# Date logic
 NOW=$(date +"%Y-%m-%d-%H%M%S")
+EPOCH=$(date "+%s")
+KEEP_ALL_DATE=$(($EPOCH - 86400))       # 1 day ago
+KEEP_DAILIES_DATE=$(($EPOCH - 2678400)) # 31 days ago
+
+export IFS=$'\n' # Better for handling spaces in filenames.
 PROFILE_FOLDER="$HOME/.rsync_tmbackup"
 LOG_FILE="$PROFILE_FOLDER/$NOW.log"
 DEST=$DEST_FOLDER/$NOW
 PREVIOUS_DEST=$(fn_find_backups | sort | tail -n 1)
 INPROGRESS_FILE=$DEST_FOLDER/backup.inprogress
-KEEP_ALL_DATE=$(date -d '-1 day' +%s)
-KEEP_DAILIES_DATE=$(date -d '-1 month' +%s)
 
 # -----------------------------------------------------------------------------
 # Create profile folder if it doesn't exist
@@ -138,7 +146,7 @@ fi
 while [ "1" ]; do
 
 	# -----------------------------------------------------------------------------
-	# Check if we are doing an incremental backup (if previous backup exists) or not
+	# Check if we are doing an incremental backup (if previous backup exists).
 	# -----------------------------------------------------------------------------
 
 	LINK_DEST_OPTION=""
@@ -165,26 +173,29 @@ while [ "1" ]; do
 	# Purge certain old backups before beginning new backup.
 	# -----------------------------------------------------------------------------
 
-	for fname in $(fn_find_backups | sort -r); do
-		date=$(basename "$fname")
-		stamp=$(fn_parse_date $date)
+	# Default value for $PREV ensures that the most recent backup is never deleted.
+	PREV="0000-00-00-000000"
+	for FILENAME in $(fn_find_backups | sort -r); do
+		BACKUP_DATE=$(basename "$FILENAME")
+		TIMESTAMP=$(fn_parse_date $BACKUP_DATE)
 
 		# Skip if failed to parse date...
-		[ -n "$stamp" ] || continue
-
-		if   [ $stamp -ge $KEEP_ALL_DATE ]; then
-			true
-
-		elif [ $stamp -ge $KEEP_DAILIES_DATE ]; then
-			# Delete all but the most recent of each day.
-			[ ${date:8:2} -eq ${prev:8:2} ] && fn_expire_backup "$fname"
-
-		else
-			# Delete all but the most recent of each month.
-			[ ${date:5:2} -eq ${prev:5:2} ] && fn_expire_backup "$fname"
+		if [ -z "$TIMESTAMP" ]; then
+			fn_log_warn "Could not parse date: $FILENAME"
+			continue
 		fi
 
-		prev=$date
+		if   [ $TIMESTAMP -ge $KEEP_ALL_DATE ]; then
+			true
+		elif [ $TIMESTAMP -ge $KEEP_DAILIES_DATE ]; then
+			# Delete all but the most recent of each day.
+			[ "${BACKUP_DATE:0:10}" == "${PREV:0:10}" ] && fn_expire_backup "$FILENAME"
+		else
+			# Delete all but the most recent of each month.
+			[ "${BACKUP_DATE:0:7}" == "${PREV:0:7}" ] && fn_expire_backup "$FILENAME"
+		fi
+
+		PREV=$BACKUP_DATE
 	done
 
 	# -----------------------------------------------------------------------------
@@ -239,39 +250,26 @@ while [ "1" ]; do
 		grep --quiet "Result too large (34)" "$LOG_FILE"
 		NO_SPACE_LEFT="$?"
 	fi
-		
-	rm -- "$LOG_FILE"
-	
-	if [ "$NO_SPACE_LEFT" == "0" ]; then
-		# TODO: -y flag
-		read -p "It looks like there is no space left on the destination. Delete old backup? (Y/n) " yn
-		case $yn in
-			[Nn]* ) exit 0;;
-		esac
 
+	rm -- "$LOG_FILE"
+
+	if [ "$NO_SPACE_LEFT" == "0" ]; then
 		fn_log_warn "No space left on device - removing oldest backup and resuming."
-		
+
 		BACKUP_FOLDER_COUNT=$(fn_find_backups | wc -l)
-		if [ "$BACKUP_FOLDER_COUNT" -lt "2" ]; then
+		if [[ "$BACKUP_FOLDER_COUNT" -lt "2" ]]; then
 			fn_log_error "No space left on device, and no old backup to delete."
 			exit 1
 		fi
-				
+
 		OLD_BACKUP_PATH=$(fn_find_backups | head -n 1)
-		if [ "$OLD_BACKUP_PATH" == "" ]; then
+		if [[ "$OLD_BACKUP_PATH" == "" ]]; then
 			fn_log_error "No space left on device, and cannot get path to oldest backup to delete."
 			exit 1
 		fi
-				
-		# Double-check that we're on a backup destination to be completely sure we're deleting the right folder
-		OLD_BACKUP_PARENT_PATH=$(dirname -- "$OLD_BACKUP_PATH")
-		if [ "$(fn_is_backup_destination $OLD_BACKUP_PARENT_PATH)" != "1" ]; then
-			fn_log_error "'$OLD_BACKUP_PATH' is not on a backup destination - aborting."
-			exit 1
-		fi
-		
+
 		fn_expire_backup "$OLD_BACKUP_PATH"
-		
+
 		# Resume backup
 		continue
 	fi
@@ -281,8 +279,14 @@ while [ "1" ]; do
 		exit $RSYNC_EXIT_CODE
 	fi
 
-	rm -f $DEST_FOLDER/latest
-	ln -sr $DEST $DEST_FOLDER/latest
+	# -----------------------------------------------------------------------------
+	# Add symlink to last successful backup
+	# -----------------------------------------------------------------------------
+
+	cd "$DEST_FOLDER"
+	rm -f -- "latest"
+	ln -s -- $(basename -- "$DEST") "latest"
+	cd -
 
 	rm -- "$INPROGRESS_FILE"
 	# TODO: grep for "^rsync error:.*$" in log
